protegendo as rotas do backend

+++ server/src/routes/memories.ts
@@ -3,8 +3,16 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+
+    app.get('/memories', async (request) => {
+
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy: {
                 createdAt: 'asc',
             }
@@ -19,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
     })
     
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -31,6 +39,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             },
         })
+
+        if (!memory.isPublic && memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
 
         return memory
     })
@@ -49,12 +61,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '4e13529c-d537-437e-8e3e-686368bfb0cb'
+                userId: request.user.sub
             }
         })
     })
     
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -69,7 +81,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+        
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -83,12 +105,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
     
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+        
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({
             where: {
